requirements.txt
azure-common==1.1.3
azure-nspkg==1.0.0
azure-storage==0.31.0
Flask==0.10.1
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.3
requests==2.9.1
six==1.10.0
Werkzeug==0.11.5
mysql-connector==2.1.4
redis==2.10.6
